requirements.txt
# FastAPI Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
